test file structure now mirrors src file structure, added more unit test coverage
