+++ test/model/replication/Replisome.test.ts
@@ -513,4 +513,105 @@ describe('Replisome', () => {
       }
     });
   });
+
+  describe('error handling and edge cases', () => {
+    test('create method handles non-Error exceptions', () => {
+      // Create a fork that will cause constructor to throw
+      const badFork = null as any; // This will cause an error in constructor
+
+      const result = Replisome.create(badFork, E_COLI);
+      expect(result.success).toBe(false);
+      if (!result.success) {
+        expect(result.error).toContain('Failed to create replisome');
+      }
+    });
+
+    test('constructor handles enzyme creation failures', () => {
+      // Mock EnzymeFactory to return failures for testing each enzyme type
+      const originalEnzymeFactory = require('../../../src/model/replication/enzyme/EnzymeFactory.js').EnzymeFactory;
+
+      // Test helicase creation failure
+      const mockEnzymeFactory = {
+        ...originalEnzymeFactory,
+        createHelicase: jest.fn().mockReturnValue({ success: false, error: 'Helicase creation failed' }),
+        createPrimase: jest.fn().mockReturnValue({ success: true, data: {} }),
+        createPolymerase: jest.fn().mockReturnValue({ success: true, data: {} }),
+        createLigase: jest.fn().mockReturnValue({ success: true, data: {} }),
+        createExonuclease: jest.fn().mockReturnValue({ success: true, data: {} }),
+      };
+
+      // Replace the module temporarily
+      const enzymeFactoryModule = require('../../../src/model/replication/enzyme/EnzymeFactory.js');
+      const originalFactory = enzymeFactoryModule.EnzymeFactory;
+      enzymeFactoryModule.EnzymeFactory = mockEnzymeFactory;
+
+      expect(() => new Replisome(fork, E_COLI)).toThrow('Helicase creation failed');
+
+      // Test primase creation failure
+      mockEnzymeFactory.createHelicase.mockReturnValue({ success: true, data: {} });
+      mockEnzymeFactory.createPrimase.mockReturnValue({ success: false, error: 'Primase creation failed' });
+
+      expect(() => new Replisome(fork, E_COLI)).toThrow('Primase creation failed');
+
+      // Test leading polymerase creation failure
+      mockEnzymeFactory.createPrimase.mockReturnValue({ success: true, data: {} });
+      mockEnzymeFactory.createPolymerase
+        .mockReturnValueOnce({ success: false, error: 'Leading polymerase creation failed' })
+        .mockReturnValue({ success: true, data: {} });
+
+      expect(() => new Replisome(fork, E_COLI)).toThrow('Leading polymerase creation failed');
+
+      // Test lagging polymerase creation failure
+      mockEnzymeFactory.createPolymerase
+        .mockReturnValueOnce({ success: true, data: {} })
+        .mockReturnValueOnce({ success: false, error: 'Lagging polymerase creation failed' });
+
+      expect(() => new Replisome(fork, E_COLI)).toThrow('Lagging polymerase creation failed');
+
+      // Test ligase creation failure
+      mockEnzymeFactory.createPolymerase.mockReturnValue({ success: true, data: {} });
+      mockEnzymeFactory.createLigase.mockReturnValue({ success: false, error: 'Ligase creation failed' });
+
+      expect(() => new Replisome(fork, E_COLI)).toThrow('Ligase creation failed');
+
+      // Test exonuclease creation failure
+      mockEnzymeFactory.createLigase.mockReturnValue({ success: true, data: {} });
+      mockEnzymeFactory.createExonuclease.mockReturnValue({ success: false, error: 'Exonuclease creation failed' });
+
+      expect(() => new Replisome(fork, E_COLI)).toThrow('Exonuclease creation failed');
+
+      // Restore original
+      enzymeFactoryModule.EnzymeFactory = originalFactory;
+    });
+
+    test('handles primer generation failure during fragment initiation', () => {
+      const replisome = new Replisome(fork, E_COLI);
+
+      // Mock RNAPrimer.generateRandom to fail
+      const originalGenerateRandom = require('../../../src/model/replication/RNAPrimer.js').RNAPrimer.generateRandom;
+      require('../../../src/model/replication/RNAPrimer.js').RNAPrimer.generateRandom = jest.fn()
+        .mockReturnValue({ success: false, error: 'Primer generation failed' });
+
+      // Try to advance the fork which should trigger fragment initiation
+      expect(() => replisome.advanceFork(1000)).toThrow('Failed to create primer: Primer generation failed');
+
+      // Restore original
+      require('../../../src/model/replication/RNAPrimer.js').RNAPrimer.generateRandom = originalGenerateRandom;
+    });
+
+    test('handles fragment creation failure during fragment initiation', () => {
+      const replisome = new Replisome(fork, E_COLI);
+
+      // Mock OkazakiFragment.create to fail
+      const originalCreate = require('../../../src/model/replication/OkazakiFragment.js').OkazakiFragment.create;
+      require('../../../src/model/replication/OkazakiFragment.js').OkazakiFragment.create = jest.fn()
+        .mockReturnValue({ success: false, error: 'Fragment creation failed' });
+
+      // Try to advance the fork which should trigger fragment initiation
+      expect(() => replisome.advanceFork(1000)).toThrow('Failed to create fragment: Fragment creation failed');
+
+      // Restore original
+      require('../../../src/model/replication/OkazakiFragment.js').OkazakiFragment.create = originalCreate;
+    });
+  });
 });
